Pivotal ID # 183230982: Avoid Special Characters In File Paths
- Refactor out the methods related to the referenced files in the persistence layer
- Fix a couple of "TooManyFunctions" warning
- Fix typos
- Change gitlab build from concurrent to sequential
