+++ submission/file-sources/src/main/kotlin/uk/ac/ebi/io/sources/SubmissionFilesSource.kt
@@ -1,6 +1,6 @@
 package uk.ac.ebi.io.sources
 
-import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
+import ac.uk.ebi.biostd.persistence.common.service.SubmissionFilesPersistenceService
 import ebi.ac.uk.extended.mapping.from.toExtAttribute
 import ebi.ac.uk.extended.model.ExtFile
 import ebi.ac.uk.extended.model.ExtSubmission
@@ -17,7 +17,7 @@ class SubmissionFilesSource(
     private val nfsFiles: PathSource,
     private val fireClient: FireClient,
     private val previousVersionFiles: Map<String, ExtFile>,
-    private val queryService: SubmissionPersistenceQueryService,
+    private val filesRepository: SubmissionFilesPersistenceService,
 ) : FilesSource {
     override val description: String
         get() = "Previous version files"
@@ -31,7 +31,7 @@ class SubmissionFilesSource(
     }
 
     private fun findSubmissionFile(path: String): ExtFile? {
-        return previousVersionFiles[path] ?: queryService.findReferencedFile(sub, path)
+        return previousVersionFiles[path] ?: filesRepository.findReferencedFile(sub, path)
     }
 
     private fun getFile(file: ExtFile): File? {
