+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/GeneralConfig.kt
@@ -1,7 +1,7 @@
 package ac.uk.ebi.biostd.common.config
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
-import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
+import ac.uk.ebi.biostd.persistence.common.service.SubmissionFilesPersistenceService
 import ac.uk.ebi.biostd.submission.domain.helpers.TempFileGenerator
 import ac.uk.ebi.biostd.submission.helpers.FilesSourceFactory
 import ac.uk.ebi.biostd.submission.service.FileSourcesService
@@ -20,8 +20,8 @@ internal class GeneralConfig {
     fun fireFilesSourceFactory(
         fireClient: FireClient,
         applicationProperties: ApplicationProperties,
-        queryService: SubmissionPersistenceQueryService,
-    ): FilesSourceFactory = FilesSourceFactory(fireClient, applicationProperties, queryService)
+        filesRepository: SubmissionFilesPersistenceService,
+    ): FilesSourceFactory = FilesSourceFactory(fireClient, applicationProperties, filesRepository)
 
     @Bean
     fun fileSourcesService(
