+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/domain/service/SubmissionQueryService.kt
@@ -4,6 +4,7 @@ import ac.uk.ebi.biostd.integration.SerializationService
 import ac.uk.ebi.biostd.integration.SubFormat
 import ac.uk.ebi.biostd.persistence.common.model.BasicSubmission
 import ac.uk.ebi.biostd.persistence.common.request.SubmissionFilter
+import ac.uk.ebi.biostd.persistence.common.service.SubmissionFilesPersistenceService
 import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
 import ebi.ac.uk.extended.mapping.to.ToFileListMapper
 import ebi.ac.uk.extended.mapping.to.ToSubmissionMapper
@@ -13,6 +14,7 @@ import java.nio.file.Files
 
 class SubmissionQueryService(
     private val submissionPersistenceQueryService: SubmissionPersistenceQueryService,
+    private val filesRepository: SubmissionFilesPersistenceService,
     private val serializationService: SerializationService,
     private val toSubmissionMapper: ToSubmissionMapper,
     private val toFileListMapper: ToFileListMapper,
@@ -27,7 +29,7 @@ class SubmissionQueryService(
 
     fun getFileList(accNo: String, fileListName: String, format: SubFormat): File {
         val submission = submissionPersistenceQueryService.getExtByAccNo(accNo)
-        val fileList = submissionPersistenceQueryService.getReferencedFiles(submission, fileListName)
+        val fileList = filesRepository.getReferencedFiles(submission, fileListName)
         val targetFile = Files.createTempFile(accNo, fileListName)
         return toFileListMapper.serialize(fileList, format, targetFile.toFile())
     }
