+++ submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/helpers/FilesSourceFactory.kt
@@ -1,7 +1,7 @@
 package ac.uk.ebi.biostd.submission.helpers
 
 import ac.uk.ebi.biostd.common.properties.ApplicationProperties
-import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceQueryService
+import ac.uk.ebi.biostd.persistence.common.service.SubmissionFilesPersistenceService
 import ebi.ac.uk.extended.model.ExtSubmission
 import ebi.ac.uk.extended.model.allInnerSubmissionFiles
 import ebi.ac.uk.io.sources.FilesSource
@@ -15,7 +15,7 @@ import java.nio.file.Paths
 class FilesSourceFactory(
     private val fireClient: FireClient,
     private val applicationProperties: ApplicationProperties,
-    private val queryService: SubmissionPersistenceQueryService,
+    private val filesRepository: SubmissionFilesPersistenceService,
 ) {
     fun createFireSource(): FilesSource = FireFilesSource(fireClient)
 
@@ -26,6 +26,6 @@ class FilesSourceFactory(
             .allInnerSubmissionFiles
             .groupBy { it.filePath }
             .mapValues { it.value.first() }
-        return SubmissionFilesSource(sub, nfsFiles, fireClient, previousVersionFiles, queryService)
+        return SubmissionFilesSource(sub, nfsFiles, fireClient, previousVersionFiles, filesRepository)
     }
 }
